Name AWS groups from the email address when using groups sync method
We're using the other sync method, "user_groups", but would like to
switch to the "groups" on because it runs a lot more efficiently, just
over 4.5 minutes compared to the "user_groups" one which is timing out
at 15 minutes near the end of the process.

Unfortunately, the "groups" method names AWS group from the email
address in Google i.e. $name@$domain.com. The more efficient method,
"groups", uses the name of the group in Google i.e. $name.

This means if we switch, it'll remove all the old groups and create new
ones, which will then remove everyone's permissions until the account
assignments are added to the new groups. This process is easy in
Terraform, but would create an outage on permissions and has the same
problem when switching back as well.

Instead, let's set the "groups" method to name them after the email as
well.

I've also removed a duplicate loop and ran the LSP formatter on the
code. Most formatting changes are whitespace so easy to ignore.